backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
asyncssh==2.14.2
python-multipart>=0.0.9
pyyaml>=6.0.1
